# Conflicts:
#	.circleci/config.yml
#	.meteor/versions
#	Dockerfile
#	client/components/boards/boardBody.js
#	client/components/boards/boardHeader.jade
#	client/components/cards/cardDetails.js
#	client/components/cards/minicard.jade
#	client/components/lists/listBody.js
#	docker-compose-build.yml
#	docker-compose.yml
#	models/boards.js
#	models/lists.js

+++ Dockerfile
@@ -82,10 +82,12 @@ ARG LOGOUT_ON_MINUTES
 ARG CORS
 ARG DEFAULT_AUTHENTICATION_METHOD
 
+ARG NO_CLEANUP
+
 # Set the environment variables (defaults where required)
 # DOES NOT WORK: paxctl fix for alpine linux: https://github.com/wekan/wekan/issues/1303
 # ENV BUILD_DEPS="paxctl"
-ENV BUILD_DEPS="apt-utils bsdtar gnupg gosu wget curl bzip2 build-essential python python3 python3-distutils git ca-certificates gcc-7" \
+ENV BUILD_DEPS="apt-utils bsdtar gnupg gosu wget curl bzip2 build-essential python python3 python3-distutils git ca-certificates gcc-7 procps locales" \
     DEBUG=false \
     NODE_VERSION=v8.15.1 \
     METEOR_RELEASE=1.6.0.1 \
@@ -302,19 +304,24 @@ RUN \
     mv /home/wekan/app_build/bundle /build && \
     \
     # Put back the original tar
-    mv $(which tar)~ $(which tar) && \
-    \
+    mv $(which tar)~ $(which tar) &&\
+
+    # RUN echo -e 'LANG=en_US.UTF-8\nLC_ALL=en_US.UTF-8' > /etc/default/locale
+
+
     # Cleanup
-    apt-get remove --purge -y ${BUILD_DEPS} && \
-    apt-get autoremove -y && \
-    npm uninstall -g api2html &&\
-    rm -R /var/lib/apt/lists/* && \
-    rm -R /home/wekan/.meteor && \
-    rm -R /home/wekan/app && \
-    rm -R /home/wekan/app_build && \
-    cat /home/wekan/python/esprima-python/files.txt | xargs rm -R && \
-    rm -R /home/wekan/python && \
-    rm /home/wekan/install_meteor.sh
+    if [ "x$NO_CLEANUP" = "x" ]; then \
+      apt-get remove --purge -y ${BUILD_DEPS} && \
+      apt-get autoremove -y && \
+      npm uninstall -g api2html &&\
+      rm -R /var/lib/apt/lists/* && \
+      rm -R /home/wekan/.meteor && \
+      rm -R /home/wekan/app && \
+      rm -R /home/wekan/app_build && \
+      cat /home/wekan/python/esprima-python/files.txt | xargs rm -R && \
+      rm -R /home/wekan/python && \
+      rm /home/wekan/install_meteor.sh; \
+    fi
 
 ENV PORT=8080
 EXPOSE $PORT
